Renamed `public` folder to `assets`
