+++ packages/components/src/assets/fonts/NeueMachina-Ultrabold.woff2
