+++ packages/components/src/assets/img/gradientLarge.webp
