+++ packages/components/src/assets/img/gradientSmall.webp
